Types change in ResourceMeter. Pair class added.

+++ org.jage.student.examples-solutions/src/main/java/org/jage/examples/resource_meter/IResourceMeterStrategy.java
@@ -4,8 +4,8 @@ import org.jage.strategy.IStrategy;
 
 public interface IResourceMeterStrategy extends IStrategy {
 
-	Long getCpuLoad();
+	Integer getCpuLoad();
 
-	Long getMemoryLoad();
+	Integer getMemoryLoad();
 
 }
